+++ docs/pbm_test_code.py
@@ -27,7 +27,8 @@ display = adafruit_ili9341.ILI9341(display_bus, width=320, height=240)
 splash = displayio.Group(max_size=10)
 display.show(splash)
 #image = "images/netpbm_p1_mono_ascii.pbm"
-image = "images/netpbm_p4_mono_binary.pbm"
+#image = "images/netpbm_p4_mono_binary.pbm"
+image = "images/netpbm_p2_ascii.pgm"
 
 bitmap, palette = adafruit_imageload.load(
     image, bitmap=displayio.Bitmap, palette=displayio.Palette
